requirements.txt
Django>=4.0.3,<4.1.0
djangorestframework>=3.13.1,<3.14.0
psycopg2-binary>=2.9.3,<2.10.0
Pillow>=9.0.1,<9.1.0
requests
graphene-django>=2.15.0,<2.16.0
python-dotenv>=0.19.2,<0.20.0
django-background-tasks>=1.2.5,<1.3.0